refactor and link nomination page
