--- src/components/Landing.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-// import FormSubmit from './form/FormSubmit';
-
-const Landing = () => (
-  <div>
-    <h1>
-      Landing Page
-    </h1>
-    {/* <FormSubmit /> */}
-  </div>
-);
-
-export default Landing;
